filters are somewhat working again w/ burnettk
